Refactor chat components for improved UI and functionality
- Updated ChatCard and ChatCardForProfile components to enhance message display and input handling.
- Implemented sticky input field and improved message scrolling behavior.
- Added loading states and disabled states for input fields and buttons.
- Removed unnecessary console logs and cleaned up code for better readability.
- Added delete confirmation modal in MyListings page for better user experience.
- Improved error handling and user feedback in various components.
- Introduced ProtectedRoute component to manage access to protected routes based on authentication status.
-Add Google Map Functionality

+++ api/controllers/deleteListing.controller.js
@@ -1,12 +1,31 @@
 import Residence from "../models/residence.model.js";
+import Amenity from "../models/amenities.model.js";
+import Lifestyle from "../models/lifestyle.model.js";
+import { v2 as cloudinary } from "cloudinary"; 
 
 export const deletePost = async (req, res) => {
   const postid = req.params.id;
   const Post = await Residence.findById(postid);
+  if (!Post) {
+    return res.status(404).json({ error: "Post Not Found" });
+  }
   console.log(Post);
   try {
-    const deletedPost = await Residence.findByIdAndDelete(postid);
+    if (Post.amenities) {
+      await Amenity.findByIdAndDelete(Post.amenities);
+    }
+    if (Post.lifestyle) {
+      await Lifestyle.findByIdAndDelete(Post.lifestyle);
+    }
 
+    if (Post.photos && Post.photos.length > 0) {
+      const deletePromises = Post.photos.map((photo) =>
+        cloudinary.uploader.destroy(photo.public_id)
+      );
+      await Promise.all(deletePromises);
+    }
+
+    const deletedPost = await Residence.findByIdAndDelete(postid);
     if (!deletedPost) {
       return res.status(404).json({ error: "Listing Not Found" });
     }
